feat: create core base for react/node/legacy lint configs
