+++ packages/eslint-plugin/src/configs/node-legacy.js
@@ -5,8 +5,8 @@ const configHelpers = require('../utils/configHelpers');
 
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: [path.join(__dirname, 'node')],
+  extends: [path.join(__dirname, 'core')],
   rules: {
-    ...configHelpers.getNamingConventionRule(true /* prefixWithI */),
+    ...configHelpers.getNamingConventionRule(true),
   },
 };
