+++ packages/eslint-plugin/src/configs/node.js
@@ -4,7 +4,7 @@ const path = require('path');
 
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: [path.join(__dirname, 'react')],
+  extends: [path.join(__dirname, 'core')],
   rules: {
     'no-console': 'off',
   },
