+++ packages/eslint-plugin/src/configs/react-legacy.js
@@ -5,10 +5,10 @@ const configHelpers = require('../utils/configHelpers');
 
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: [path.join(__dirname, 'react')],
+  extends: [path.join(__dirname, 'core'), path.join(__dirname, 'react-config')],
 
   rules: {
-    ...configHelpers.getNamingConventionRule(true /* prefixWithI */),
+    ...configHelpers.getNamingConventionRule(true),
     'jsdoc/check-tag-names': 'off',
   },
 };
